requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
apscheduler>=3.10.0
email-validator>=2.0.0
python-dotenv>=1.0.0
